Fix another debug warning
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@9625 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ src/ll/HWTopoLinux.cpp
@@ -350,7 +350,7 @@ struct AutoLinuxPolicy {
 
     //Compute first thread in package
     leaders.resize(numPackages, -1);
-    for (int i = 0; i < packages.size(); ++i)
+    for (int i = 0; i < (int) packages.size(); ++i)
       if (leaders[packages[i]] == -1)
 	leaders[packages[i]] = i;
